Merged PR 72: Global Notes Work Items 880-886
Completed the modal. Did not get the dynamic modal working and did not get the named router outlet for dialog working. There is still the positioning issue on the side nav for the sub menu. I can't figure which styles are governing that and they are overriding whatever I try to do.

Related work items: #692, #880, #882, #883, #884, #885